+++ crates/cayenne/src/ddl/mod.rs
@@ -14,17 +14,22 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-//! Single-node Cayenne DDL: handler, operations, and physical execution plans.
+//! Single-node Cayenne DDL: handler, operations, physical execution plans,
+//! logical nodes, and local MERGE planner.
 //!
 //! Usable without the runtime crate. Pair with `datafusion_ddl::DdlAnalyzerRule`
 //! and `datafusion_ddl::DdlExtensionPlanner` for full DDL support in a standalone
 //! Cayenne-backed `SessionContext`.
 
 pub mod handler;
+pub mod logical_nodes;
+pub mod merge_planner;
 pub mod operations;
 pub mod physical_plans;
 
 pub use handler::CayenneDdlHandler;
+pub use logical_nodes::CayenneMergeNode;
+pub use merge_planner::CayenneDmlExtensionPlanner;
 pub use physical_plans::CayenneMergeExec;
 
 use datafusion::catalog::CatalogProvider;
